# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
...
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}:
    put:
      tags:
      - team-controller
      operationId: update_7
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDto"
...
  /api/v1/platform/clubs/{id}/reactivate:
    post:
      tags:
      - club-controller
      operationId: reactivate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClubDto"
...
  /api/v1/manage/clubs/{clubId}/sponsors/{sponsorId}/reactivate:
    post:
      tags:
      - sponsor-controller
      operationId: reactivate_1
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SponsorDto"
  /api/v1/manage/clubs/{clubId}/sponsors/{sponsorId}/deactivate:
    post:
      tags:
      - sponsor-controller
      operationId: deactivate
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SponsorDto"
...
  /api/v1/manage/clubs/{clubId}/sponsors/{sponsorId}/contacts/{contactId}/reactivate:
    post:
      tags:
      - sponsor-contact-controller
      operationId: reactivate_2
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SponsorContactDto"
  /api/v1/manage/clubs/{clubId}/sponsors/{sponsorId}/contacts/{contactId}/deactivate:
    post:
      tags:
      - sponsor-contact-controller
      operationId: deactivate_1
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SponsorContactDto"
...
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams:
    get:
      tags:
      - team-controller
      operationId: listBySection
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamDto"
    post:
      tags:
      - team-controller
      operationId: create_6
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
        required: true
      responses:
        "201":
          description: Team created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/sponsors/{sponsorId}/unlink:
    post:
      tags:
      - team-controller
      operationId: unlinkSponsor
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/sponsors/{sponsorId}/link:
    post:
      tags:
      - team-controller
      operationId: linkSponsor
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sponsorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/reactivate:
    post:
      tags:
      - team-controller
      operationId: reactivate_3
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/deactivate:
    post:
      tags:
      - team-controller
      operationId: deactivate_2
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/contacts/{contactId}/unlink:
    post:
      tags:
      - team-controller
      operationId: unlinkContact
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/contacts/{contactId}/link:
    post:
      tags:
      - team-controller
      operationId: linkContact
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkTeamContactRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/reactivate:
    post:
      tags:
      - section-controller
      operationId: reactivate_4
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SectionDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/deactivate:
    post:
      tags:
      - section-controller
      operationId: deactivate_3
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SectionDto"
...
  /api/v1/manage/clubs/{clubId}/seasons/{seasonId}/reactivate:
    post:
      tags:
      - season-controller
      operationId: reactivate_5
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SeasonDto"
  /api/v1/manage/clubs/{clubId}/seasons/{seasonId}/deactivate:
    post:
      tags:
      - season-controller
      operationId: deactivate_4
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SeasonDto"
...
  /api/v1/manage/clubs/{clubId}/players/{playerId}/reactivate:
    post:
      tags:
      - player-controller
      operationId: reactivate_6
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlayerDto"
  /api/v1/manage/clubs/{clubId}/players/{playerId}/deactivate:
    post:
      tags:
      - player-controller
      operationId: deactivate_5
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlayerDto"
...
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/reactivate:
    post:
      tags:
      - match-controller
      operationId: reactivate_7
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchDto"
...
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/deactivate:
    post:
      tags:
      - match-controller
      operationId: deactivate_6
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchDto"
...
  /api/v1/manage/clubs/{clubId}/leagues/{leagueId}/reactivate:
    post:
      tags:
      - league-controller
      operationId: reactivate_8
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: leagueId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeagueDto"
  /api/v1/manage/clubs/{clubId}/leagues/{leagueId}/deactivate:
    post:
      tags:
      - league-controller
      operationId: deactivate_7
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: leagueId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeagueDto"
...
  /api/v1/manage/clubs/{clubId}/contacts/{contactId}/reactivate:
    post:
      tags:
      - club-contact-controller
      operationId: reactivate_9
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClubContactDto"
  /api/v1/manage/clubs/{clubId}/contacts/{contactId}/deactivate:
    post:
      tags:
      - club-contact-controller
      operationId: deactivate_8
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: contactId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClubContactDto"
...
  /api/v1/manage/clubs/{clubId}/teams:
    get:
      tags:
      - team-controller
      operationId: listByClub
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamDto"
...
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/sponsors:
    get:
      tags:
      - team-controller
      operationId: listSponsors
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SponsorDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/teams/{teamId}/contacts:
    get:
      tags:
      - team-controller
      operationId: listContacts
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamContactDto"
  /api/v1/manage/clubs/{clubId}/sections/{sectionId}/contacts:
    get:
      tags:
      - section-controller
      operationId: listContacts_1
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClubContactDto"
...
components:
  schemas:
...
    ClubDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum:
          - ONBOARDING
          - ACTIVE
          - SUSPENDED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    SocialLinkDto:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
          pattern: ^https?://.+
      required:
      - platform
      - url
...
    SponsorDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLinkDto"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    SponsorContactDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sponsorId:
          type: string
          format: uuid
        contact:
          $ref: "#/components/schemas/ContactDto"
        role:
          type: string
        isPrimary:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    SectionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        parentSectionId:
          type: string
          format: uuid
        name:
          type: string
        minAge:
          type: integer
          format: int32
        maxAge:
          type: integer
          format: int32
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
      required:
      - name
    TeamDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        sectionId:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    SeasonDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        label:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    PlayerDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        photoUrl:
          type: string
        clubMembershipNumber:
          type: string
        medicalAidProvider:
          type: string
        medicalAidMemberNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        altContactName:
          type: string
        altContactPhone:
          type: string
        battingStance:
          type: string
          enum:
          - RIGHT_HANDED
          - LEFT_HANDED
        bowlingArm:
          type: string
          enum:
          - RIGHT_ARM
          - LEFT_ARM
        bowlingType:
          type: string
          enum:
          - FAST
          - FAST_MEDIUM
          - MEDIUM_FAST
          - MEDIUM
          - OFF_BREAK
          - LEG_BREAK
          - ORTHODOX_SPIN
          - WRIST_SPIN
          - GOOGLY
        isWicketKeeper:
          type: boolean
        active:
          type: boolean
        sectionIds:
          type: array
          items:
            type: string
            format: uuid
        jerseyNumber:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    MatchDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        homeTeamId:
          type: string
          format: uuid
        homeTeamName:
          type: string
        awayTeamId:
          type: string
          format: uuid
        awayTeamName:
          type: string
        leagueId:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        matchDate:
          type: string
          format: date-time
        venue:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
      required:
      - name
    LinkTeamContactRequest:
      type: object
      properties:
        role:
          type: string
      required:
      - role
...
    TeamContactDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contact:
          $ref: "#/components/schemas/ClubContactDto"
        role:
          type: string
        createdAt:
          type: string
          format: date-time
...